feat: create artisan shortcut and add functionality to dashboard view
'php artisan' can be invoked by typing 'a', users can be added in
dashboard view and user specs are displayed in the dashboard
(pagination needs to be fixed). movies form currently non-functional,
corresponding CRUD methods need to be implemented in MovieController.
Review CRUD is not implemented, only a ReviewController skeleton is.